master

+++ code/pool/sqlconnpool.cpp
@@ -6,7 +6,7 @@ SqlConnPool* SqlConnPool::Instance() {
 }
 
 // 初始化
-void SqlConnPool::Init(const char* host, int port,
+void SqlConnPool::Init(const char* host, uint16_t port,
               const char* user,const char* pwd, 
               const char* dbName, int connSize = 10) {
     assert(connSize > 0);
